Refactor API and improve services

+++ TBC.Task.API/Controllers/PersonsController.cs
@@ -1,10 +1,11 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
 using AutoMapper;
+using TBC.Task.API.Localization;
 using TBC.Task.API.Models;
+using TBC.Task.API.Resources;
 using TBC.Task.Domain;
 using TBC.Task.Domain.Interfaces.Services;
-using TBC.Task.API.Resources;
 
 namespace TBC.Task.API.Controllers;
 
@@ -14,25 +15,28 @@ public class PersonsController : ControllerBase
 {
 	private readonly IPersonService _personService;
 	private readonly IRelatedPersonService _relatedPersonService;
-	private readonly IMapper _mapper;
 	private readonly IHostEnvironment _environment;
 	private readonly IStringLocalizer<ErrorResources> _errorLocalizer;
+	private readonly IMapper _mapper;
+	private readonly IConfiguration _configuration;
 	private readonly ILogger<PersonsController> _logger;
 
 	public PersonsController(
 		IPersonService personService,
 		IRelatedPersonService relatedPersonService,
-		IMapper mapper,
-		IHostEnvironment environment, 
+		IHostEnvironment environment,
 		IStringLocalizer<ErrorResources> errorLocalizer,
+		IMapper mapper,
+		IConfiguration configuration,
 		ILogger<PersonsController> logger)
 	{
 		_personService = personService ?? throw new ArgumentNullException(nameof(personService));
 		_relatedPersonService = relatedPersonService ?? throw new ArgumentNullException(nameof(relatedPersonService));
-		_mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
 		_environment = environment ?? throw new ArgumentNullException(nameof(environment));
-		_logger = logger ?? throw new ArgumentNullException(nameof(logger));
+		_mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
 		_errorLocalizer = errorLocalizer ?? throw new ArgumentNullException(nameof(errorLocalizer));
+		_logger = logger ?? throw new ArgumentNullException(nameof(logger));
+		_configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
 	}
 
 	[HttpPost]
@@ -47,6 +51,9 @@ public class PersonsController : ControllerBase
 	[HttpPut]
 	public async Task<IActionResult> Update(int id, RequestPersonModel model)
 	{
+		if (!_personService.Exists(id))
+			return NotFound(_errorLocalizer.GetLocalized(ErrorResources.PersonNotFound));
+
 		var person = _mapper.Map<Person>(model);
 		person.Id = id;
 		_personService.Update(person);
@@ -57,34 +64,44 @@ public class PersonsController : ControllerBase
 	[HttpDelete("{id:int}")]
 	public async Task<IActionResult> Delete(int id)
 	{
+		if (!_personService.Exists(id))
+			return NotFound(_errorLocalizer.GetLocalized(ErrorResources.PersonNotFound));
+
 		_personService.Delete(id);
 
 		return NoContent();
+	}
+
+	[HttpGet("{id:int}")]
+	public async Task<IActionResult> Get(int id)
+	{
+		var person = BuildGetResponse(id);
+		if (person == null)
+			return NotFound(_errorLocalizer.GetLocalized(ErrorResources.PersonNotFound));
+
+		return Ok(person);
 	}
 
 	[HttpPost]
 	[Route("AddRelatedPerson/{from:int}/{to:int}")]
 	public async Task<IActionResult> AddRelatedPerson(int from, int to)
 	{
+		if (_relatedPersonService.Exists(from, to))
+			return Ok();
+
 		_relatedPersonService.Insert(new RelatedPerson { FromId = from, ToId = to });
 
-		return NoContent();
+		return Ok();
 	}
 
 	[HttpDelete]
 	[Route("DeleteRelatedPerson/{from:int}/{to:int}")]
 	public async Task<IActionResult> DeleteRelatedPerson(int from, int to)
 	{
-		var relatedPerson = _relatedPersonService
-			.Set(x => x.FromId == from && x.ToId == to)
-			.FirstOrDefault();
+		if (!_relatedPersonService.Exists(from, to))
+			return NotFound(_errorLocalizer.GetLocalized(ErrorResources.RelationNotFound));
 
-		if (relatedPerson == null)
-		{
-			return NotFound();
-		}
-
-		_relatedPersonService.Delete(relatedPerson);
+		_relatedPersonService.Delete(from, to);
 
 		return NoContent();
 	}
@@ -93,50 +110,52 @@ public class PersonsController : ControllerBase
 	[Route("RelatedPersonsCount/{id:int}")]
 	public async Task<IActionResult> GetRelatedPersonsCount(int id)
 	{
-		var relatedPersonsCount = _relatedPersonService.GetRelatedPersonsCount(id);
+		if (!_personService.Exists(id))
+			return NotFound(_errorLocalizer.GetLocalized(ErrorResources.PersonNotFound));
 
-		return Ok(new { relatedPersonsCount });
+		return Ok(new
+		{
+			relatedPersonsCount = _relatedPersonService.GetRelatedPersonsCount(id)
+		});
 	}
 
-	[HttpGet("{id:int}")]
-	public async Task<IActionResult> Get(int id)
-	{
-		var person = BuildGetResponse(id);
-
-		return Ok(person);
-	}
-	
-	[HttpPost]
+	[HttpPatch]
 	[Route("UploadPhoto")]
 	public async Task<IActionResult> UploadPhoto(int id, IFormFile file)
 	{
+		if (!_personService.Exists(id))
+			return NotFound(_errorLocalizer.GetLocalized(ErrorResources.PersonNotFound));
 		if (file.Length == 0)
-		{
-			return Ok();
-		}
-		
+			return BadRequest();
+
 		var person = _personService.Get(id);
 		var imageFilePath = SavePhoto(id, file);
 
 		person.PhotoPath = imageFilePath;
-		person.PhotoUrl = $"/Persons/Photo/{person.Id}";
-		
+		person.PhotoUrl = $"{_configuration["PhotoRelativeUrl"]}/{person.Id}";
+
 		_personService.Update(person);
 
-		return NoContent();
+		return Ok(new
+		{
+			id,
+			photoSize = file.Length
+		});
 	}
 
 	[HttpGet]
 	[Route("Photo/{id:int}")]
 	public async Task<IActionResult> GetPhoto(int id)
 	{
-		var person = _personService.Get(id);
-		if (string.IsNullOrEmpty(person.PhotoPath))
-		{
-			return NoContent();
-		}
+		if (!_personService.Exists(id))
+			return NotFound(_errorLocalizer.GetLocalized(ErrorResources.PersonNotFound));
 
-		var data = await System.IO.File.ReadAllBytesAsync(person.PhotoPath);         
+		var person = _personService.Get(id);
+		var data = await GetPhotoData(person);
+
+		if (!data.Any())
+			return NotFound();
+
 		return File(data, "image/jpeg");
 	}
 
@@ -146,9 +165,16 @@ public class PersonsController : ControllerBase
 	{
 		var result = _personService
 			.QuickSearch(keyword, currentPage, pageSize)
-			.Select(p => _mapper.Map<ResponsePersonModel>(p));
+			.Select(p => _mapper.Map<ResponsePersonModel>(p))
+			.ToArray();
 
-		return Ok(result);
+		return Ok(new
+		{
+			currentPage,
+			pageSize,
+			totalPages = (int) Math.Ceiling((double) result.Count() / pageSize),
+			result
+		});
 	}
 
 	[HttpGet]
@@ -156,17 +182,27 @@ public class PersonsController : ControllerBase
 	public async Task<ActionResult<IEnumerable<Person>>> Search(string keyword, int currentPage, int pageSize = 10)
 	{
 		var result = _personService
-			.QuickSearch(keyword, currentPage, pageSize)
-			.Select(p => _mapper.Map<ResponsePersonModel>(p));
+			.Search(keyword, currentPage, pageSize)
+			.Select(p => _mapper.Map<ResponsePersonModel>(p))
+			.ToArray();
 
-		return Ok(result);
+		return Ok(new
+		{
+			currentPage,
+			pageSize,
+			totalPages = (int) Math.Ceiling((double) result.Count() / pageSize),
+			result
+		});
 	}
 
 	#region Private helper methods
 
-	private ResponsePersonWithRelatedModel BuildGetResponse(int id)
+	private ResponsePersonWithRelatedModel? BuildGetResponse(int id)
 	{
 		var person = _personService.GetIncludeCity(id);
+		if (person == null)
+			return null;
+
 		var relatedPersons = _relatedPersonService.GetRelatedPersons(id);
 
 		var response = _mapper.Map<ResponsePersonWithRelatedModel>(person);
@@ -175,15 +211,21 @@ public class PersonsController : ControllerBase
 		return response;
 	}
 
+	private static async Task<byte[]> GetPhotoData(Person person)
+	{
+		if (string.IsNullOrEmpty(person.PhotoPath) || !System.IO.File.Exists(person.PhotoPath))
+			return Array.Empty<byte>();
+
+		return await System.IO.File.ReadAllBytesAsync(person.PhotoPath);
+	}
+
 	private string SavePhoto(int id, IFormFile file)
 	{
 		var imageDirectory = Path.Combine(_environment.ContentRootPath, "Uploads", "Photos", id.ToString());
 		var imageFilePath = Path.Combine(imageDirectory, file.FileName);
 
 		if (!Directory.Exists(imageDirectory))
-		{
 			Directory.CreateDirectory(imageDirectory);
-		}
 
 		using var fileStream = System.IO.File.Create(Path.Combine(imageFilePath));
 		file.CopyTo(fileStream);
